Updated JS and HTML to filter properly.
Debugged the filtering JS to properly create the Fusion Tables
queries.
Currently the filtering works if you hit the exact, but
it is not an ideal user scenario, as Fusion Tables doesn't
allow ORs.
This means the best way to use these filters is to instead
make multiple queries and correlate them.

Work done with tons of help from RCH (Thanks!)