line up data in popups
Fixes #6

Let me know what you think or how I can improve it.

• Updated layout of map info window see: https://f.cloud.github.com/assets/1526861/262091/5c142c64-8d15-11e2-8670-fb5a42119897.png
• Added a underscore template to index.html ( I think there is a little to much logic in it but might be a ok first pass )
• Fixed displaying multiple instances of the info window when clicking the same pin.
• Added closing other info windows if open.

My jshint is not set up to handle ASI and is all like >:( on save. I'm not 1337 enough yet to use it either so I added a few semicolons.
If thats a problem I can fix that.